spawnbox's spawn and block rotation
- now we have prototype of spawnbox
- one more command "setrotation"
- sketch of game boss

+++ Tests/Tests/utils/Vectors.cs
@@ -26,7 +26,7 @@ namespace Tests.utils
             //Act
             //Rotating to U-turn
             //Assert
-            Assert.AreEqual(vector.Rotate(Math.PI), expected);
+            Assert.AreEqual(vector.Rotate((float) Math.PI), expected);
         }
 
         [Test]
